# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aigoni API
  version: 1.0.0
  description: |
    Aigoni API contract. `x-implementation-status` marks routes registered now
    implemented REST contract.
servers:
  - url: /
...
paths:
...
  /api/admin/v1/tags:
    get:
      tags: [Admin API]
      summary: List post tag candidates
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      responses:
        '200':
          description: Tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
...
  /api/admin/v1/wiki/chat:
    post:
      tags: [Admin API]
      summary: Start one natural-language Wiki Agent run
      description: Admin Wiki Chat and machine Wiki Ask share the `AIGONI_WIKI_AGENT_CONCURRENCY` active-run pool. Pending and running runs count toward the limit; exceeding it returns 409.
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WikiChatRequest'}
      responses:
        '202': {description: Agent run accepted, content: {application/json: {schema: {$ref: '#/components/schemas/WikiRun'}}}}
        '409': {$ref: '#/components/responses/Error'}
  /api/admin/v1/wiki/chat:stream:
    post:
      tags: [Admin API]
      summary: Start one Wiki Agent run and stream its events
      description: Response Content-Type is `text/event-stream`. Events are `step`, `tool`, `delta`, `done`, and `error`. The backend run is not tied to the SSE connection lifecycle. Admin Wiki Chat and machine Wiki Ask share the `AIGONI_WIKI_AGENT_CONCURRENCY` active-run pool.
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WikiChatRequest'}
      responses:
        '200': {description: SSE stream, content: {text/event-stream: {schema: {type: string}}}}
        '409': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: aigoni_session
      description: HttpOnly, SameSite=Lax; Secure over HTTPS.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on Admin write operations in the target contract.
    APIKeyHeader: {type: apiKey, in: header, name: X-API-Key}
    BearerAuth: {type: http, scheme: bearer}
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
...
    Tag: {type: object, required: [name, count], properties: {name: {type: string}, count: {type: integer}}}
    TagList: {type: array, items: {$ref: '#/components/schemas/Tag'}}
...
    WikiRun: {type: object, required: [id, kind, status], properties: {id: {type: string}, kind: {type: string}, status: {type: string, enum: [pending, running, completed, failed]}, message: {type: string}, question: {type: string}, answer: {type: string}, answer_html: {type: string}, reasoning: {type: string}, sources: {type: array, items: {type: string}}, files: {type: array, items: {type: string}}, events: {type: array, items: {$ref: '#/components/schemas/WikiRunEvent'}}, error: {type: string}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    WikiRunEvent: {type: object, required: [seq, event, payload], properties: {seq: {type: integer}, time: {type: string, format: date-time}, event: {type: string}, payload: {type: object}}}
    WikiChatRequest:
      type: object
      required: [message]
      properties:
        message: {type: string}
        history:
          type: array
          items:
            type: object
            properties:
              role: {type: string, enum: [user, assistant]}
              content: {type: string}
...
    ErrorResponse: {type: object, required: [error], properties: {error: {oneOf: [{type: string}, {$ref: '#/components/schemas/ErrorDetail'}]}}}
    ErrorDetail: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}, field_errors: {type: object, additionalProperties: {type: string}}, request_id: {type: string}}}